Change the write_graph_lgl test back into example
