--- tests/unit/igraph_write_graph_lgl.out
@@ -1,37 +1,0 @@
-Test without isolates:
-# 0
-1
-2
-# 1
-2
-3
-# 2
-4
-# 3
-4
-
-Test with isolates:
-# 0
-1
-2
-# 1
-2
-3
-# 2
-4
-# 3
-4
-# 5
-# 6
-
-Test vertex and edge labels:
-# zero
-one 8
-two 9
-# one
-two 10
-three 11
-# two
-four 12
-# three
-four 13
